feat: accomplish docs' structure
