+++ docs/index.md
@@ -1,1 +1,7 @@
-## 开始
+---
+title: 介绍
+---
+
+# @mlz/admin
+
+[@mlz/admin](https://github.com/juicecube/mlz-admin) 是按照`编程猫设计规范`，基于[Antd](https://ant.design/index-cn) 修订的组件库。旨在提供统一的后台系统设计和交互模型，以及大幅提高开发效率。
