+++ docs/others/iris.md
@@ -1,0 +1,5 @@
+# Iris
+
+> 内部账号系统 jssdk
+
+使用`@mlz-pack`或`yummi`托管项目，比如@mlz/pack：
