+++ docs/start/compatibility.md
@@ -1,0 +1,7 @@
+## 兼容性
+
+兼容全部`Antd^4.1.5`的组件和 API，在实际开发中可作为 Antd 的替代库。除此之外，加入一些符合编程猫业务逻辑和规范的特有组件和 API。
+
+> ps1. 本文档文案和展示均面向编程猫设计规范。所以没有出现在本文档的并不代表没有，记住是**完全兼容**的。
+>
+> ps2. 新加的 Apis 均展示在每个组件的 APIs 段落
