+++ docs/start/design-std.md
@@ -1,0 +1,1 @@
+# 编程猫设计规范
