+++ docs/start/installation.md
@@ -1,0 +1,17 @@
+---
+title: 安装
+order: 5
+---
+
+# 安装
+
+使用 npm 安装`@mlz/admin`：
+
+```shell
+npm install @mlz/admin --save
+```
+
+在 webpack 中配置 less-loader：
+
+```js
+```
