+++ src/shared/components/keep-alive/keep-alive.type.ts
@@ -1,0 +1,10 @@
+// / <reference types="react" />
+import React from 'react';
+
+export interface KeepAliveProps {
+  children: React.ReactElement;
+  name: string;
+  timeout?: number;
+  onLoad?: () => void;
+  props: any;
+}
